move high level parser to HighLevelParser
